force conversion size_t->uint64_t for osx compatibility, fixing #39

+++ fflas-ffpack/fflas/fflas_helpers.inl
@@ -58,12 +58,12 @@ namespace FFLAS {
 		template <> inline size_t min_types(double& k) {return static_cast<size_t>(std::min(k,double(std::numeric_limits<size_t>::max())));}
 		template <> inline size_t min_types(int64_t& k) {return static_cast<size_t>(std::min(k,int64_t(std::numeric_limits<size_t>::max())));}
 #endif
-		template <> inline size_t min_types(RecInt::rint<6>& k) {return static_cast<size_t>(std::min(k,RecInt::rint<6>(std::numeric_limits<size_t>::max())));}
-		template <> inline size_t min_types(RecInt::rint<7>& k) {return static_cast<size_t>(std::min(k,RecInt::rint<7>(std::numeric_limits<size_t>::max())));}
-		template <> inline size_t min_types(RecInt::rint<8>& k) {return static_cast<size_t>(std::min(k,RecInt::rint<8>(std::numeric_limits<size_t>::max())));}
-		template <> inline size_t min_types(RecInt::rint<9>& k) {return static_cast<size_t>(std::min(k,RecInt::rint<9>(std::numeric_limits<size_t>::max())));}
-		template <> inline size_t min_types(RecInt::rint<10>& k) {return static_cast<size_t>(std::min(k,RecInt::rint<10>(std::numeric_limits<size_t>::max())));}
-		template <> inline size_t min_types(Givaro::Integer& k) {return static_cast<size_t>(std::min(k,Givaro::Integer(std::numeric_limits<size_t>::max())));}
+		template <> inline size_t min_types(RecInt::rint<6>& k) {return static_cast<size_t>(std::min(k,RecInt::rint<6>(uint64_t(std::numeric_limits<size_t>::max()))));}
+		template <> inline size_t min_types(RecInt::rint<7>& k) {return static_cast<size_t>(std::min(k,RecInt::rint<7>(uint64_t(std::numeric_limits<size_t>::max()))));}
+		template <> inline size_t min_types(RecInt::rint<8>& k) {return static_cast<size_t>(std::min(k,RecInt::rint<8>(uint64_t(std::numeric_limits<size_t>::max()))));}
+		template <> inline size_t min_types(RecInt::rint<9>& k) {return static_cast<size_t>(std::min(k,RecInt::rint<9>(uint64_t(std::numeric_limits<size_t>::max()))));}
+		template <> inline size_t min_types(RecInt::rint<10>& k) {return static_cast<size_t>(std::min(k,RecInt::rint<10>(uint64_t(std::numeric_limits<size_t>::max()))));}
+		template <> inline size_t min_types(Givaro::Integer& k) {return static_cast<size_t>(std::min(k,Givaro::Integer(uint64_t(std::numeric_limits<size_t>::max()))));}
 
 		template <class T>
 		inline bool unfit(T x){return false;}
